requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pymongo
motor
bson